+++ src/test-utils.cpp
@@ -140,3 +140,56 @@ context("euclidean distance") {
     expect_equal_matrix(euclidean(coord), expected);
   }
 }
+
+context("convert_eigen and convert_tmb") {
+  test_that("convert_eigen and convert_tmb are reversible") {
+    matrix<double> tmb_mat(4, 2);
+    tmb_mat << 1, 2, 3, 4, 5, 6, 7, 8;
+    expect_equal_matrix(convert_eigen(convert_tmb(tmb_mat)), tmb_mat);
+  }
+}
+
+context("cpow works") {
+  test_that("cpow gives correct power by element") {
+    matrix<double> tmb_mat(4, 2);
+    tmb_mat << 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0;
+    matrix<double> expected(4, 2);
+    expected << 1.0, sqrt(2.0), sqrt(3.0), 2.0, sqrt(5.0), sqrt(6.0), sqrt(7.0), sqrt(8.0);
+    expect_equal_matrix(convert_eigen(cpow(convert_tmb(tmb_mat), 0.5)), expected);
+  }
+  test_that("cpow gives correct power by element") {
+    matrix<double> tmb_mat(4, 2);
+    tmb_mat << 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0;
+    matrix<double> expected(4, 2);
+    expected << 1.0, 4.0, 9.0, 16.0, 25.0, 36.0, 49.0, 64.0;
+    expect_equal_matrix(convert_eigen(cpow(convert_tmb(tmb_mat), 2.0)), expected);
+  }
+}
+
+context("pseudoInverseSqrt works") {
+  test_that("pseudoInverseSqrt gives correct result(compare to R)") {
+    matrix<double> tmb_mat(3, 3);
+    tmb_mat << 5.483417,  2.861011,  3.478399,
+      2.861011,  3.169936, -1.075550,
+      3.478399, -1.075550, 10.525825;
+    
+    matrix<double> expected(3, 3);
+    expected << 0.8235633, -0.5514385, -0.2586037,
+      -0.5514385,  1.0568775,  0.2548210,
+      -0.2586037,  0.2548210,  0.4095994;
+    expect_equal_matrix(pseudoInverseSqrt(tmb_mat), expected);
+  }
+
+  test_that("pseudoInverseSqrt gives correct result for rank-deficient matrix") {
+    matrix<double> tmb_mat(3, 3);
+    tmb_mat << 5.483417,  2.861011,  0,
+      2.861011,  3.169936, 0,
+      0, 0, 0;
+    
+    matrix<double> expected(3, 3);
+    expected << 0.5331152, -0.2459070,    0.0,
+      -0.2459070, 0.7319613,    0.0,
+      0.0000000,  0.0000000,    0.0;
+    expect_equal_matrix(pseudoInverseSqrt(tmb_mat), expected);
+  }
+}